style: :lipstick: change status icon colours
So "done" is the lightest green, "in progress" is darker", and both "planned" and "potential" are grey. I think this shows more clearly what we are currently working on (green nuances) and what are in the future (grey). These colours also come from the colour scale we got from Mjølner.