style: :lipstick: change status icon colours (#73)
## Description

So "done" is the lightest green, "in progress" is darker", and both
"planned" and "potential" are grey. I think this shows more clearly what
we are currently working on (green nuances) and what are in the future
(grey). These colours also come from the colour scale we got from
Mjølner.

### Before 

![image](https://github.com/user-attachments/assets/2e86e2ca-759c-4197-911c-5564c9329a92)


### Now 

![image](https://github.com/user-attachments/assets/757df974-de93-4d58-b0f6-f944d0d87bd0)


Closes #

## Reviewer Focus

<!-- Please delete as appropriate: -->
This PR needs an in-depth review.

## Checklist

- [X] Rendered website locally