docs: :memo: add flashtalk transcript for the Open Science Seminar in June, 2025 (#130)
# Description

I've made a transcript for what I plan to present in the flashtalk
tomorrow at the Open Science Seminar.

If you have the time, I would love some feedback 😁 

I've practiced it (a.k.a. read it out loud) and it's just under 3
minutes.

This PR needs an in-depth review.

## Checklist

- [X] Formatted Markdown
- [X] Ran `just run-all`

---------

Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>